Moved Mailgun message ID handling to shared services
ref https://linear.app/ghost/issue/BER-3851

Gift delivery also needs to interpret Mailgun send results, so keeping the helper under automations would make a cross-service transport concern appear automation-owned. Moving it to the shared services area gives existing and future mail callers one neutral implementation without changing behavior.
