+++ ghost/core/core/server/services/email-analytics/automation-email-analytics-batch-processor.ts
@@ -1,4 +1,4 @@
-import {normalizeMailgunMessageId} from '../automations/mailgun-message-id';
+import {normalizeMailgunMessageId} from '../lib/mailgun-message-id';
 import type * as automationsApi from '../automations/automations-api';
 import type {AutomatedEmailEvents, AutomatedEmailRecipientWithMailgunId} from '../automations/automations-repository';
 import type {BatchEventProcessor} from './batch-event-processor';
